#321 preparing for unit testing
